+++ src/carpadi_market/filters.py
@@ -5,17 +5,19 @@ from src.models.models import CarProduct, CarTransmissionTypes, FuelTypes, CarTy
 
 
 class CarProductFilter(filters.FilterSet):
-    selling_price = filters.NumericRangeFilter(field_name="selling_price")
+    selling_price = filters.NumberFilter(field_name="selling_price", lookup_expr="lte")
     search = filters.CharFilter(field_name="car", method="search_field")
-    year = filters.DateFromToRangeFilter(field_name="car__year")
+    year = filters.NumberFilter(field_name="car__information__brand__year")
     age = filters.NumericRangeFilter(field_name="car__age")
     transmission = filters.ChoiceFilter(
-        field_name="car__transmission", lookup_expr="iexact", choices=CarTransmissionTypes.choices
+        field_name="car__information__transmission", lookup_expr="iexact", choices=CarTransmissionTypes.choices
     )
-    fuel_type = filters.ChoiceFilter(field_name="car__fuel_type", lookup_expr="iexact", choices=FuelTypes.choices)
-    car_type = filters.ChoiceFilter(field_name="car__car_type", lookup_expr="iexact", choices=CarTypes.choices)
+    fuel_type = filters.ChoiceFilter(field_name="car__information__fuel_type", lookup_expr="iexact", choices=FuelTypes.choices)
+    car_type = filters.CharFilter(field_name="car__information__car_type", lookup_expr="icontains")
+    make = filters.CharFilter(field_name="car__information__manufacturer", lookup_expr="icontains")
 
     def search_field(self, queryset, name, value):
+
         return queryset.filter(
             Q(car__information__brand__model__icontains=value)
             | Q(car__information__manufacturer__icontains=value)
@@ -28,7 +30,4 @@ class CarProductFilter(filters.FilterSet):
 
     class Meta:
         model = CarProduct
-        fields = [
-            "selling_price",
-            "status",
-        ]
+        fields = ["selling_price", "status", 'year', 'transmission', 'search', 'fuel_type', 'car_type', 'id']